cnbackend/pip-requirements.txt
django=4.0.6
djangorestframework=3.13.1
djangorestframework-jwt=1.11.0